Add policy fixture on test environment
Summary: Tests didn't work because the dnssec_parent_authority variable was pointing to a parent authority that was only created when running tests with the RAILS_ENV environment variable set to 'test'. Setting this variable to 'test', leads to the dnssec_policies table to be empty, meaning that we need fixtures for the test database.

Test Plan: none

Reviewers: #servers

Subscribers: #servers

Differential Revision: https://repo.grnet.gr/D44